dev: pmic: Change data type for pwm_value from uint8_t to uint32_t
Avoid overflow by using acceptable data type uint32_t for pwm_value,
which is a 9-bit PWM register value.

CRs-Fixed: 611459

+++ dev/pmic/pm8921/include/dev/pm8921_pwm.h
@@ -151,7 +151,7 @@ struct pm8921_pwm_config {
     uint8_t clk;
     uint8_t pre_div;
     uint8_t pre_div_exp;
-    uint8_t pwm_value;
+    uint32_t pwm_value;
     uint8_t bypass_lut;
     uint8_t pwm_ctl[NUM_LPG_CTL_REGS];
 };
